Portfolio no.2
Theres lottsss of ways i can make this website look better, and hopefully more in line with who i am as a person. I think im still getting used to the tricks, and systems of this but once i get the basics all the way down, ill be good to go!